+++ native/opencv-wrapper.h
@@ -41,9 +41,6 @@ void c_drop(void* value);
 //   Highgui: high-level GUI
 // =============================================================================
 
-void cv_imshow(const char* const winname, cv::Mat* mat);
-int cv_wait_key(int delay_in_millis);
-
 // =============================================================================
 //
 // =============================================================================
